[refactor/lang-infra] Cleanup unwrappers

+++ src/main/kotlin/icu/windea/pls/cwt/codeInsight/unwrap/CwtPropertyRemover.kt
@@ -2,19 +2,23 @@ package icu.windea.pls.cwt.codeInsight.unwrap
 
 import com.intellij.psi.PsiElement
 import icu.windea.pls.ChronicleBundle
+import icu.windea.pls.core.util.values.or
+import icu.windea.pls.core.util.values.unresolved
 import icu.windea.pls.cwt.psi.CwtProperty
 
 class CwtPropertyRemover : CwtUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val name = if (e is CwtProperty) e.name else ""
-        return ChronicleBundle.message("cwt.remove.property", name)
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is CwtProperty
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is CwtProperty
+    override fun getDescription(element: PsiElement): String {
+        if (element !is CwtProperty) return "" // unexpected
+        val name = element.name
+        return ChronicleBundle.message("cwt.remove.property", name.or.unresolved())
     }
 
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is CwtProperty) return // unexpected
         context.delete(element)
     }
 }
