+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/unwrap/ParadoxLocalisationCommandRemover.kt
@@ -5,16 +5,18 @@ import icu.windea.pls.ChronicleBundle
 import icu.windea.pls.localisation.psi.ParadoxLocalisationCommand
 
 class ParadoxLocalisationCommandRemover : ParadoxLocalisationUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val text = if (e is ParadoxLocalisationCommand) e.text else ""
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is ParadoxLocalisationCommand
+    }
+
+    override fun getDescription(element: PsiElement): String {
+        if (element !is ParadoxLocalisationCommand) return "" // unexpected
+        val text = element.presentableText
         return ChronicleBundle.message("localisation.remove.command", text)
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is ParadoxLocalisationCommand
-    }
-
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is ParadoxLocalisationCommand) return // unexpected
         context.delete(element)
     }
 }
