+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/unwrap/ParadoxLocalisationConceptCommandRemover.kt
@@ -5,16 +5,18 @@ import icu.windea.pls.ChronicleBundle
 import icu.windea.pls.localisation.psi.ParadoxLocalisationConceptCommand
 
 class ParadoxLocalisationConceptCommandRemover : ParadoxLocalisationUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val text = if (e is ParadoxLocalisationConceptCommand) e.text else ""
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is ParadoxLocalisationConceptCommand
+    }
+
+    override fun getDescription(element: PsiElement): String {
+        if (element !is ParadoxLocalisationConceptCommand) return "" // unexpected
+        val text = element.presentableText
         return ChronicleBundle.message("localisation.remove.conceptCommand", text)
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is ParadoxLocalisationConceptCommand
-    }
-
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is ParadoxLocalisationConceptCommand) return // unexpected
         context.delete(element)
     }
 }
