+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/unwrap/ParadoxLocalisationIconRemover.kt
@@ -2,19 +2,23 @@ package icu.windea.pls.localisation.codeInsight.unwrap
 
 import com.intellij.psi.PsiElement
 import icu.windea.pls.ChronicleBundle
+import icu.windea.pls.core.util.values.or
+import icu.windea.pls.core.util.values.unresolved
 import icu.windea.pls.localisation.psi.ParadoxLocalisationIcon
 
 class ParadoxLocalisationIconRemover : ParadoxLocalisationUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val name = if (e is ParadoxLocalisationIcon) e.name.orEmpty() else ""
-        return ChronicleBundle.message("localisation.remove.icon", name)
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is ParadoxLocalisationIcon
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is ParadoxLocalisationIcon
+    override fun getDescription(element: PsiElement): String {
+        if (element !is ParadoxLocalisationIcon) return "" // unexpected
+        val name = element.name
+        return ChronicleBundle.message("localisation.remove.icon", name.or.unresolved())
     }
 
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is ParadoxLocalisationIcon) return // unexpected
         context.delete(element)
     }
 }
