+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/unwrap/ParadoxLocalisationPropertyRemover.kt
@@ -2,19 +2,23 @@ package icu.windea.pls.localisation.codeInsight.unwrap
 
 import com.intellij.psi.PsiElement
 import icu.windea.pls.ChronicleBundle
+import icu.windea.pls.core.util.values.or
+import icu.windea.pls.core.util.values.unresolved
 import icu.windea.pls.localisation.psi.ParadoxLocalisationProperty
 
 class ParadoxLocalisationPropertyRemover : ParadoxLocalisationUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val name = if (e is ParadoxLocalisationProperty) e.name else ""
-        return ChronicleBundle.message("localisation.remove.property", name)
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is ParadoxLocalisationProperty
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is ParadoxLocalisationProperty
+    override fun getDescription(element: PsiElement): String {
+        if (element !is ParadoxLocalisationProperty) return "" // unexpected
+        val name = element.name
+        return ChronicleBundle.message("localisation.remove.property", name.or.unresolved())
     }
 
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is ParadoxLocalisationProperty) return // unexpected
         context.delete(element)
     }
 }
