+++ src/main/kotlin/icu/windea/pls/script/codeInsight/unwrap/ParadoxScriptConditionalBlockUnwrapper.kt
@@ -2,21 +2,21 @@ package icu.windea.pls.script.codeInsight.unwrap
 
 import com.intellij.psi.PsiElement
 import icu.windea.pls.ChronicleBundle
-import icu.windea.pls.core.util.values.or
-import icu.windea.pls.core.util.values.unresolved
-import icu.windea.pls.script.psi.ParadoxScriptConditionalBlock
+import icu.windea.pls.script.psi.ParadoxScriptNormalConditionalBlock
 
 class ParadoxScriptConditionalBlockUnwrapper : ParadoxScriptUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val name = if (e is ParadoxScriptConditionalBlock) e.presentableText.or.unresolved() else ""
-        return ChronicleBundle.message("script.unwrap.conditionalBlock", name)
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is ParadoxScriptNormalConditionalBlock
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is ParadoxScriptConditionalBlock
+    override fun getDescription(element: PsiElement): String {
+        if (element !is ParadoxScriptNormalConditionalBlock) return "" // unexpected
+        val text = element.presentableText
+        return ChronicleBundle.message("script.unwrap.conditionalBlock", text)
     }
 
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is ParadoxScriptNormalConditionalBlock) return // unexpected
         context.extract(element, element)
         context.delete(element)
     }
