+++ src/main/kotlin/icu/windea/pls/script/codeInsight/unwrap/ParadoxScriptInlineConditionalBlockUnwrapper.kt
@@ -2,21 +2,21 @@ package icu.windea.pls.script.codeInsight.unwrap
 
 import com.intellij.psi.PsiElement
 import icu.windea.pls.ChronicleBundle
-import icu.windea.pls.core.util.values.or
-import icu.windea.pls.core.util.values.unresolved
 import icu.windea.pls.script.psi.ParadoxScriptInlineConditionalBlock
 
 class ParadoxScriptInlineConditionalBlockUnwrapper : ParadoxScriptUnwrapper() {
-    override fun getDescription(e: PsiElement): String {
-        val name = if (e is ParadoxScriptInlineConditionalBlock) e.presentableText.or.unresolved() else ""
-        return ChronicleBundle.message("script.unwrap.inlineConditionalBlock", name)
+    override fun isApplicableTo(element: PsiElement): Boolean {
+        return element is ParadoxScriptInlineConditionalBlock
     }
 
-    override fun isApplicableTo(e: PsiElement): Boolean {
-        return e is ParadoxScriptInlineConditionalBlock
+    override fun getDescription(element: PsiElement): String {
+        if (element !is ParadoxScriptInlineConditionalBlock) return "" // unexpected
+        val text = element.presentableText
+        return ChronicleBundle.message("script.unwrap.inlineConditionalBlock", text)
     }
 
     override fun doUnwrap(element: PsiElement, context: Context) {
+        if (element !is ParadoxScriptInlineConditionalBlock) return // unexpected
         context.extract(element, element)
         context.delete(element)
     }
